% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{acs5.2011.county}
\alias{acs5.2011.county}
\title{County Level American Community Survey 5-Year Data from 2013}
\usage{
acs5.2011.county(token, state = "*", county = "*", variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{variables}{Vector of variable codes.}
}
\description{
County Level American Community Survey 5-Year Data from 2013
}
\examples{
example = acs5.2011.county(token, state = c("47", "53"), county = "*", variables)
example = acs5.2011.county(token, variables =  variables)
head(example)
state county B01003_001E B02001_002E B02001_003E
1    01    001       53944       42577        9755
2    01    003      179523      155068       16936
3    01    005       27546       13576       12632
4    01    007       22746       17437        5153
5    01    009       57140       54446         806
6    01    011       10877        3115        7619
}
