% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{acs5.2011.state}
\alias{acs5.2011.state}
\title{State Level American Community Survey 5-Year Data from 2013}
\usage{
acs5.2011.state(token, state = "*", variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{variables}{Vector of variable codes.}
}
\description{
State Level American Community Survey 5-Year Data from 2013
}
\examples{
acs5.2011.state(token, state = c("47", "53"), variables = c("B01003_001E", "B02001_002E", "B02001_003E"))
}
