% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{acs5.2013.blockgroup}
\alias{acs5.2013.blockgroup}
\title{Block-Group Level American Community Survey 5-Year Data from 2013}
\usage{
acs5.2013.blockgroup(token, state = "*", variables)
}
\arguments{
\item{token}{Go to http://www.census.gov/developers/ to request an API key. This function will not work if you do not have your own unique key. There is no default.}

\item{state}{A state FIPS code (stored as a character, ie "47"). Defaults to "*" for every state in the US.}

\item{variables}{The desired variables for which you would like data. Variable names found here: http://api.census.gov/data/2011/acs5/variables.html}
}
\description{
Block-Group Level American Community Survey 5-Year Data from 2013
}
