% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blockgroup.2000}
\alias{blockgroup.2000}
\title{2000 blockgroup-level data}
\usage{
blockgroup.2000(token, state = "*", county = "*", blockgroup = "*",
  variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{blockgroup}{Vector of blockgroup numbers. Defaults to "*" for all blockgroup}

\item{variables}{Vector of variable codes.}
}
\description{
2000 blockgroup-level data
}
\examples{
blockgroup.2000(token = token, state = c(47, 48), county = "*", blockgroup = "*", variables = c("P001001","P001001"))
blockgroup.2000(token = token, state = 47, county = 1, blockgroup = "*", variables = c("P001001","P001001"))
}

