% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blockgroup.data}
\alias{blockgroup.data}
\title{Blockgroup-level Decennial Census data}
\usage{
blockgroup.data(token, state = "*", county = "*", blockgroup = "*",
  variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{blockgroup}{Vector of blockgroup numbers. Defaults to "*" for all blockgroup}

\item{variables}{Vector of variable codes.}
}
\description{
Blockgroup-level Decennial Census data
}
\examples{
blockgroup.data(token = token, state = c(47, 48), county = "*", blockgroup = "*", variables = c("P0010001","P0010001"))
blockgroup.data(token = token, state = 47, county = 1, blockgroup = "*", variables = c("P001001","P001001"), year = 2000, survey = "sf3")
}

