% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{census.2000.bg}
\alias{census.2000.bg}
\title{2000 Block Group Data (deprecated)}
\usage{
census.2000.bg(token, state = "*", variables)
}
\arguments{
\item{token}{Go to http://www.census.gov/developers/ to request an API key. This function will not work if you do not have your own unique key. There is no default.}

\item{state}{A state FIPS code (stored as a character, ie "47"). Defaults to "*" for every state in the US.}

\item{variables}{The desired variables for which you would like data. Variable names found here: http://api.census.gov/data/2000/sf3/variables.html}
}
\description{
2000 Block Group Data (deprecated)
}
\examples{
example = census.2000.bg(token = "yourkey", state = "47", variables = "P001001")
head(example)
P001001        GEOID
1    1035 470010201001
2     751 470010201002
3     577 470010201003
4     970 470010202001
5    3519 470010202002
6     714 470010202003
}
