% Generated by roxygen2 (4.0.1): do not edit by hand
\name{census.geocoder}
\alias{census.geocoder}
\title{Geocodes street addresses to latitude, longitude, and 2010 Census fips code (GEOID).}
\usage{
census.geocoder(id, street, city, state, zip, year)
}
\arguments{
\item{id}{Name of each address or row in the data.frame containing addresses}

\item{street}{Street address excluding appartment or unit numbers}

\item{city}{City name}

\item{state}{State abbreviation, ie "TN" vs "Tennessee"}

\item{zip}{Zip code. First five characters are taken}
}
\description{
Geocodes street addresses to latitude, longitude, and 2010 Census fips code (GEOID).
}
\details{
There are NO DEFAULTS set.
}
\examples{
census.geocoder("Cloyne", "2600 Ridge Road", "Berkeley", "CA", "94709", 2010)
}

