% Generated by roxygen2 (4.0.1): do not edit by hand
\name{county.2000}
\alias{county.2000}
\title{2000 county-level data}
\usage{
county.2000(token, state = "*", county = "*", variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{variables}{Vector of variable codes.}
}
\description{
2000 county-level data
}
\examples{
county.2000(token = token, state = c(47:49), county = "*", variables = c("P001001", "PCT014003"))
county.2000(token = token, state = c(47), county = 5,variables = c("P001001", "PCT014003","P001001"))
}

