% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{county.data}
\alias{county.data}
\title{County-level Decennial Census data}
\usage{
county.data(token, state = "*", county = "*", variables, year = 2010,
  survey = "sf1")
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{variables}{Vector of variable codes.}

\item{year}{Either 2000 or 2010; 1990 is partially supported but may have errors.}

\item{survey}{Either "sf1" or "sf3"}
}
\description{
County-level Decennial Census data
}
\examples{
county.data(token = token, state = c(47:49), county = "*", variables = c("P001001", "PCT014003"))
county.data(token = token, state = c(47), county = 5,variables = c("P001001", "PCT014003","P001001"))
}
