% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demogromatics-package.r
 \docType{package}
 \name{demogromatics}
 \alias{demogromatics-package}
\title{Census Data Aquisition and Mapping}
\description{
}
\usage{
state.data(token, state = "*", variables, year = 2010, survey = "sf1")
county.data(token, state = "*", county = "*", variables, year = 2010, survey = "sf1")
tract.data(token, state = "*", county = "*", variables, year = 2010, survey = "sf1")
blockgroup.data(token, state = "*", county = "*", blockgroup = "*", variables, year = 2010, survey = "sf1")
FIPS.find(lat, long, year)
process.shapefiles(folder, shapefile)
geo.circle(center = c(0, 0), r = 100, n = 100)
download.shapefiles(url)
census.geocoder(id, street, city, state, zip, year)
census.2000.bg(token, state = "*", variables)
process.shapefiles(folder, shapefile)
mapquest.geocoder(id, street, city, state, zip, key, year)
process.api.data(datalist)
 }