% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download.shapefiles}
\alias{download.shapefiles}
\title{Batch downloads, unzips, and processes TIGER shapefiles.}
\usage{
download.shapefiles(url)
}
\arguments{
\item{url}{URL to download page for TIGER shapefiles.}
}
\description{
Batch downloads, unzips, and processes TIGER shapefiles. Unprocessed files are stored in a temporary directory and deleted post facto.
}
\examples{
TN = download.shapefiles(url = "http://www2.census.gov/geo/tiger/TIGER2010/BG/2000/tl_2010_47_bg00.zip")
head(TN)
      long      lat group        GEOID
1 -88.21252 34.99558   0.1 470719806003
2 -88.21373 34.99558   0.1 470719806003
3 -88.21489 34.99558   0.1 470719806003
4 -88.21767 34.99557   0.1 470719806003
5 -88.22340 34.99554   0.1 470719806003
6 -88.22428 34.99554   0.1 470719806003
}

