% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geo.circle}
\alias{geo.circle}
\title{Draw Circles on Maps}
\usage{
geo.circle(center = c(0, 0), r = 100, n = 100)
}
\arguments{
\item{center}{Coordinates of the center point. Defaults to c(0,0).}

\item{r}{Radius in miles. Defaults to 100 miles.}

\item{n}{Number of points to define the circle. Defaults to 100.}
}
\description{

}
\examples{
states = Cs(tennessee, kentucky, mississippi, alabama, georgia)
tn = map_data("county", region = states)
Y = mean(tn[tn$subregion == "davidson",]$lat)
X = mean(tn[tn$subregion == "davidson",]$long)
around.nashville = circle(center = c(X,Y), r = 100)
around.nashville = data.frame(around.nashville, group = rep(1, dim(around.nashville)[1]))
ggplot(tn, aes(x = long, y = lat, group = group)) +
geom_polygon(aes(fill = group),color = NA) +
 coord_map() +
 geom_point(aes(X,Y)) +
 geom_path(dat = around.nashville, aes(x = x, y = y, group = group)) +
 geom_polygon(dat = around.nashville, aes(x = x, y = y, group = group), alpha = 1/5, fill = "hotpink")
}

