% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{mapquest.geocoder}
\alias{mapquest.geocoder}
\title{Mapquest Geocoder
Geocodes street addresses to latitude, longitude, and desired year's FIPS code (GEOID).}
\usage{
mapquest.geocoder(id, street, city, state, zip, key, year)
}
\arguments{
\item{id}{Name of each address or row in the data.frame containing addresses}

\item{street}{Street address excluding appartment or unit numbers}

\item{city}{City name}

\item{state}{State abbreviation, ie "TN" vs "Tennessee"}

\item{zip}{Zip code. First five characters are taken}
}
\description{
Mapquest Geocoder
Geocodes street addresses to latitude, longitude, and desired year's FIPS code (GEOID).
}
\details{
There are NO DEFAULTS set.
}
\examples{
mapquest.geocoder("Cloyne", "2600 Ridge Road", "Berkeley", "CA", "94709", key = YOURKEY, year = 2010)

}
