% Generated by roxygen2 (4.0.1): do not edit by hand
\name{process.api.data}
\alias{process.api.data}
\title{Process API data}
\usage{
process.api.data(datalist)
}
\arguments{
\item{datalist}{The resulting list from using fromJSON()}
}
\description{
Puts json data into a usable format. Useful for extracting Census data.
}
\examples{
process.api.data(fromJSON(file=url(paste("http://api.census.gov/data/2000/sf3?key=", token,"&get=P001001&for=county:*&in=state:*", sep = ""))))
}

