% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{process.shapefiles}
\alias{process.shapefiles}
\title{Process TIGER shapefiles
Process shapefiles into a usable format. This function is helpful if you have already downloaded the shapefiles and need to process them from data stored locally. Alternatively, you can use this function as part of an API download.}
\usage{
process.shapefiles(folder, shapefile)
}
\arguments{
\item{folder}{The path of the folder that contains the shapefile and other files with .dbf, .prj, .xml, and .shx extensions.}

\item{shapefile}{The name of the shapefile WITHOUT the path or .shp extension}
}
\description{
Process TIGER shapefiles
Process shapefiles into a usable format. This function is helpful if you have already downloaded the shapefiles and need to process them from data stored locally. Alternatively, you can use this function as part of an API download.
}
\examples{
process.shapefiles(folder = "./my_shapefiles", shapefile = "tl_2010_47_county00")
process.shapefiles(folder = "./my_shapefiles", shapefile = "tl_2010_53_bg10")
}
