% Generated by roxygen2 (4.0.1): do not edit by hand
\name{state.2000}
\alias{state.2000}
\title{2000 state-level data}
\usage{
state.2000(token, state = "*", variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{variables}{Vector of variable codes.}
}
\description{
2000 state-level data
}
\examples{
state.2000(token = token, state = c(47:49), variables = c("P001001", "PCT014003"))
state.2000(token = token, variables = c("P001001", "PCT014003"))
state.2000(token = token, state = "*", variables = c("P001001", "PCT014003"))
}

