% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{state.data}
\alias{state.data}
\title{State-level Decennial Census data}
\usage{
state.data(token, state = "*", variables, year = 2010, survey = "sf1")
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{variables}{Vector of variable codes.}

\item{year}{Either 2000 or 2010; 1990 is partially supported but may have errors.}

\item{survey}{Either "sf1" or "sf3"}
}
\description{
State-level Decennial Census data
}
\examples{
state.data(token = token, state = c(47:49), variables = c("P001001", "PCT014003"))
state.data(token = token, variables = c("P001001", "PCT014003"))
state.data(token = token, state = "*", variables = c("P001001", "PCT014003"))
}
