% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tract.2000}
\alias{tract.2000}
\title{2000 tract-level data}
\usage{
tract.2000(token, state = "*", county = "*", variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{variables}{Vector of variable codes.}
}
\description{
2000 tract-level data
}
\examples{
tract.2000(token = token, state = 47, county = "*", variables = c("P001001", "PCT014003"))
tract.2000(token = token, state = 47, county = 5, variables = c("P001001", "PCT014003"))
}

