% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{blockgroup.data}
\alias{blockgroup.data}
\title{Blockgroup-level Decennial Census data}
\usage{
blockgroup.data(token, state = "*", county = "*", blockgroup = "*",
  variables, year = 2010, survey = "sf1")
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{blockgroup}{Vector of blockgroup numbers. Defaults to "*" for all blockgroup}

\item{variables}{Vector of variable codes.}

\item{year}{Either 2000 or 2010; 1990 is partially supported but may have errors.}

\item{survey}{Either "sf1" or "sf3"}
}
\description{
Blockgroup-level Decennial Census data
}
\examples{
blockgroup.data(token = token, state = c(47, 48), county = "*", blockgroup = "*", variables = c("P001001","P001001"))
blockgroup.data(token = token, state = 47, county = 1, blockgroup = "*", variables = c("P001001","P001001"))
}

