% Generated by roxygen2 (4.0.1): do not edit by hand
\name{county.data}
\alias{county.data}
\title{County-level Decennial Census data}
\usage{
county.data(token, state = "*", county = "*", variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{variables}{Vector of variable codes.}
}
\description{
County-level Decennial Census data
}
\examples{
county.data(token = token, state = c(47:49), county = "*", variables = c("P001001", "PCT014003"), year = 2000, survey = "sf3")
county.data(token = token, state = c(47), county = 5,variables = c("P0010001"), year = 2010, survey = "sf1")
county.data(token = token, state = c(47), county = "*",variables = c("P0010001"), year = 2010, survey = "sf1")
county.data(token = token, state = c(47), county = "*",variables = c("P0010001"), year = 1990, survey = "sf1")
}

