% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{demogromatics}
\alias{demogromatics}
\alias{demogromatics-package}
\title{demogromatics}
\description{
Useful functions for using US Census Bureau shapefiles and data.
}
\usage{
download.shapefiles(url)
geocode.2010(id, street, city, state, zip)
process.api.data(datalist)
process.shapefiles(folder, shapefile)
geo.circle(center = c(0,0), r = 100, n = 100)
}
\author{Nancy Organ}
