% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{demogromatics}
\alias{demogromatics}
\alias{demogromatics-package}
\title{Cesus Data Aquisition and Mapping}
\description{
Demogromatics
}
\usage{
state.2000(token, state = "*", variables)
county.2000(token, state = "*", county = "*", variables)
tract.2000(token, state = "*", county = "*", variables)
blockgroup.2000(token, state = "*", county = "*", blockgroup = "*", variables)
FIPS.find(lat, long, year)
process.shapefiles(folder, shapefile)
geo.circle(center = c(0, 0), r = 100, n = 100)
download.shapefiles(url)
census.geocoder(id, street, city, state, zip, year)
census.2000.bg(token, state = "*", variables)
process.shapefiles(folder, shapefile)
mapquest.geocoder(id, street, city, state, zip, key, year)
process.api.data(datalist)
}

 