% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demogromatics-package.r
\docType{package}
\name{demogromatics} 
\alias{demogromatics-package}
\title{Census Data Aquisition and Mapping}
\description{
Thank you for using demogromatics! If you use this package for research or publication purposes please
cite this package and the author, Nancy Organ. 
}
\usage{
state.data(token, state = "*", variables, year = 2010, survey = "sf1")
county.data(token, state = "*", county = "*", variables, year = 2010, survey = "sf1")
tract.data(token, state = "*", county = "*", variables, year = 2010, survey = "sf1")
blockgroup.data(token, state = "*", county = "*", blockgroup = "*", variables, year = 2010, survey = "sf1")
FIPS.find(lat, long, year)
process.shapefiles(folder, shapefile)
download.shapefiles(url)
census.geocoder(id, street, city, state, zip, year)
census.2000.bg(token, state = "*", variables)
process.shapefiles(folder, shapefile)
mapquest.geocoder(id, street, city, state, zip, key, year)
process.api.data(datalist)
SES.index(addresses, mapquest.key, census.key)
 }
