% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{demogromatics}
\alias{demogromatics}
\alias{demogromatics-package}
\title{demogromatics}
\description{
Process and display Census data
}
\usage{
census.2000.bg(token, state = "*", variables)download.shapefiles()
download.shapefiles(url)
FIPS.find(lat, long, year)
geo.circle(center = c(0, 0), r = 100, n = 100)
geocode.2010(id, street, city, state, zip)
mapquest.geocoder(id, street, city, state, zip, year)
process.api.data(datalist)
process.shapefiles(folder, shapefile)

}
\author{Nancy Organ}