% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FIPS.find}
\alias{FIPS.find}
\title{Finds FIPS code for a given latitude and longitude}
\usage{
FIPS.find(lat, long, year)
}
\arguments{
\item{lat}{Vector of latitudes}

\item{long}{Vector of longitudes}

\item{year}{Census year, ie. 2000, 2010, etc.}
}
\description{
Finds FIPS code for a given latitude and longitude
}
\examples{
FIPS.find(lat = 37.87622, long = -122.2585, year = 2000)
[1] "060014225002"
}

