% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{process.api.data}
\alias{process.api.data}
\title{Puts json data into a usable format.}
\usage{
process.api.data(datalist)
}
\arguments{
\item{datalist}{The resulting list from using fromJSON()}
}
\description{
Puts json data into a usable format.
}
\examples{
process.api.data(fromJSON(file=url(paste("http://api.census.gov/data/2000/sf3?key=", token,"&get=P001001&for=county:*&in=state:*", sep = ""))))

}
