% Generated by roxygen2 (4.0.1): do not edit by hand
\name{process.shapefiles}
\alias{process.shapefiles}
\title{Process TIGER shapefiles}
\usage{
process.shapefiles(folder, shapefile)
}
\arguments{
\item{folder}{The path of the folder that contains the shapefile and other files with .dbf, .prj, .xml, and .shx extensions.}

\item{shapefile}{The name of the shapefile WITHOUT the path or .shp extension}
}
\description{Process shapefiles into a usable format. This function is helpful if you have already downloaded the shapefiles and need to process them from data stored locally. Alternatively, you can use this function as part of an API download.
}
\examples{
process.shapefiles(folder = "./my_shapefiles", shapefile = "tl_2010_47_county00")
}

