% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogromatics-functions.R
\name{SES.index}
\alias{SES.index}
\title{Blockgroup-level Decennial Census data}
\usage{
SES.index(addresses, mapquest.key, census.key)
}
\arguments{
\item{addresses}{File path to a .csv or data frame of addresses}

\item{mapquest.key}{You can get a key by creating a developper's account at http://developer.mapquest.com/web/info/account/app-keys}

\item{census.key}{You can get a key from http://www.census.gov/developers/}
}
\description{
Blockgroup-level Decennial Census data
}
