% Generated by roxygen2 (4.0.1): do not edit by hand
\name{state.data}
\alias{state.data}
\title{State-level Decennial Census data}
\usage{
state.data(token, state = "*", variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{variables}{Vector of variable codes.}

\item{year}{Year of survey. Defaults to 2010.}

\item{survey}{Survey name, either "sf1" or "sf3". Defaults to "sf1".}
}
\description{
State-level Decennial Census data
}
\examples{
state.data(token = token, state = c(47:49), variables = c("P001001", "PCT014003"), year = 2000, survey = "sf1")
state.data(token = token, state = c(47:49), variables = "P0010001", year = 2010)
state.data(token = token, state = c(47:49), variables = "P0010001", year = 1990, survey = "sf3")
state.data(token = token, variables = c("P0010001"))
state.data(token = token, state = "*", variables = c("P0010001"))
state.data(token = token, variables = c("P001001"), year = 2000, survey = "sf3")
state.data(token = token, state = 45:47, variables = c("P001001"), year = 2000, survey = "sf3")

}

