% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tract.data}
\alias{tract.data}
\title{Tract-level Decennial Census data}
\usage{
tract.data(token, state = "*", county = "*", variables)
}
\arguments{
\item{token}{Unique API token}

\item{state}{Vector of state numbers. Defaults to "*" for all states.}

\item{county}{Vector of county numbers. Defaults to "*" for all counties.}

\item{variables}{Vector of variable codes.}

\item{year}{Year of survey. Defaults to 2010.}

\item{survey}{Survey name, either "sf1" or "sf3". Defaults to "sf1".}
}
\description{
Tract-level Decennial Census data
}
\examples{
tract.data(token = token, state = 47, county = "*", variables = c("P001001", "PCT014003"), year = 2000, survey = "sf3")
tract.data(token = token, state = 47, county = "*", variables = c("P0010001"))
tract.data(token = token, state = "*", county = "*", variables = "P0010001")
tract.data(token = token, state = "*", county = "*", variables = "P001001", year = 2000, survey = "sf3")
tract.data(token = token, state = 47, county = 5, variables = c("P0010001"))
}

